+++ internal/kernel/ports/usecases.go
@@ -9,3 +9,7 @@ type CreateUnitUsecase interface {
 type CreateVersionUsecase interface {
 	CreateVersion(req CreateVersionRequest) (CreateVersionResponse, error)
 }
+
+type SetMeaningUsecase interface {
+	SetMeaning(req SetMeaningRequest) (SetMeaningResponse, error)
+}